feat(frontend): Add ability to create repo
I've changed how this project is going to work. Before, it was just
going to be a bunch of disparate scripts. Now, I intend this to scaffold
an entire project in a consistent way. We should still be able to use
most if not all scripts individually, but there will be an overarching
one to scaffold a whole project at the beginning, setting up:

  * Repository
  * Organisation Project Board
  * Labels
  * GitHub Issue Templates
  * GitHub Actions

So far I've implemented the `repository` and `labels` (labels were alreay
available).y